Split shader input handling into its own module
There are now two inputs, not one per frame-buffer. We only really need
one to write to and one for the GPU to process.

Also, there is now one command-buffer per input, so that we can set up
the next frame before the previous one is finished.

The new Duo module combines the inputs and command-buffers.